Time to typescript for real, boys!
